修改附件修改

+++ expert/utils.py
@@ -224,10 +224,11 @@ def copy_img(url, identity, img_type, ecode, creater):
                 os.makedirs(file_formal_path)
             formal_file = shutil.copyfile(url, file_path)
             path = os.path.join(identity, tcode, ecode) + '/'
-            AttachmentFileinfo.objects.filter(ecode=ecode, tcode=tcode, file_name=file_name).delete()
-            AttachmentFileinfo.objects.create(ecode=ecode, tcode=tcode, file_format=1, file_name=file_name,
-                                              state=1, publish=1, file_order=0, operation_state=3,
-                                              creater=creater, path=path, file_caption=file_name)
+            if not img_type != 'consultEditor':
+                AttachmentFileinfo.objects.filter(ecode=ecode, tcode=tcode, file_name=file_name).delete()
+                AttachmentFileinfo.objects.create(ecode=ecode, tcode=tcode, file_format=1, file_name=file_name,
+                                                  state=1, publish=1, file_order=0, operation_state=3,
+                                                  creater=creater, path=path, file_caption=file_name)
             return formal_file
         return None
     except Exception as e:
